Revert "feat(Core/Entities): Enabled loading basic stats for players, pets and items from database as 32 bit integers."
This reverts commit c976141d271b2e45a6f5ea17c6d726fe7e9e19fe.

+++ src/server/game/Entities/Player/PlayerStorage.cpp
@@ -7797,7 +7797,7 @@ void Player::_SaveStats(CharacterDatabaseTransaction trans)
     stmt->SetData(index++, GetUInt32Value(UNIT_FIELD_ATTACK_POWER));
     stmt->SetData(index++, GetUInt32Value(UNIT_FIELD_RANGED_ATTACK_POWER));
     stmt->SetData(index++, GetBaseSpellPowerBonus());
-    stmt->SetData(index++, GetUInt32Value(PLAYER_FIELD_COMBAT_RATING_1 + CR_CRIT_TAKEN_SPELL));
+    stmt->SetData(index++, GetUInt32Value(PLAYER_FIELD_COMBAT_RATING_1 + static_cast<uint16>(CR_CRIT_TAKEN_SPELL)));
 
     trans->Append(stmt);
 }
